refactor: DashBoardColumn 컴포넌트 jsx 리팩토링
