+++ components/AppLayout/DashboardLayout/components/DashBoardColumn.tsx
@@ -1,12 +1,9 @@
-import { CreateTodoButton, DashBoardCard, CircleChip } from '@/components'
-import Image from 'next/image'
-import settingIcon from '@/public/icons/settingIcon.svg'
+import { CreateTodoButton, DashBoardCard } from '@/components'
 import { getDashBoardCard } from '@/service/cards'
 import { useCallback, useEffect, useState } from 'react'
 import useAsync from '@/hooks/useAsync'
 import { Card } from '@/types/dashboard'
-import { useAppDispatch } from '@/hooks/useApp'
-import { openModal } from '@/store/reducers/modalReducer'
+import DashBoardColumnHeader from './DashBoardColumnHeader'
 
 interface DashBoardColumnProps {
   columnId: number
@@ -14,8 +11,6 @@ interface DashBoardColumnProps {
 }
 
 export default function DashBoardColumn({ columnId, columnTitle }: DashBoardColumnProps) {
-  const dispatch = useAppDispatch()
-
   const [cardList, setCardList] = useState<Card[]>([])
   const { requestFunction: getCardsRequest } = useAsync(getDashBoardCard)
 
@@ -38,43 +33,14 @@ export default function DashBoardColumn({ columnId, columnTitle }: DashBoardColu
 
   return (
     <section className="flex flex-col gap-[1.6rem] pr-[2rem]">
-      <div className="mb-[0.9rem] flex items-center justify-between">
-        <div className="flex items-center gap-[0.8rem]">
-          <CircleChip color="#5534DA" />
-          <h3 className="mr-[0.4rem] text-[1.8rem] font-[700]">{columnTitle}</h3>
-          {cardList && (
-            <div className="rounded-[0.4rem] bg-var-gray2 px-[0.6rem] py-[0.3rem] text-[1.2rem] text-var-gray5">
-              {cardList.length}
-            </div>
-          )}
-        </div>
-        <button
-          type="button"
-          onClick={() =>
-            dispatch(openModal({ modalName: 'EditColumn', modalProps: { columnId, columnTitle } }))
-          }
-        >
-          <Image src={settingIcon} alt="설정" width="24" height="24" />
-        </button>
-      </div>
+      <DashBoardColumnHeader cardList={cardList} columnId={columnId} columnTitle={columnTitle} />
       <CreateTodoButton />
       {cardList.length > 0 &&
         cardList.map((cardItem) => (
           <DashBoardCard
             key={cardItem.id}
             columnTitle={columnTitle}
-            card={{
-              id: cardItem.id,
-              title: cardItem.title,
-              description: cardItem.description,
-              tags: cardItem.tags,
-              dueDate: cardItem.dueDate,
-              assignee: cardItem.assignee,
-              teamId: cardItem.teamId,
-              columnId: cardItem.columnId,
-              createdAt: cardItem.createdAt,
-              updatedAt: cardItem.updatedAt,
-            }}
+            card={cardItem}
             onDelete={handleDeleteCard}
           />
         ))}
